+++ Lithium/src/Script/MonoServer.h
@@ -34,9 +34,9 @@ namespace Lithium
 
 			//-----internal-calls------------------------------------------------------
 			static void Log(MonoString* log);
-			static bool HasComponent_Interal(int entityID,MonoObject* type);
-			static void SetPosition_Internal(int entityID, glm::vec3* vector);
-			static void GetPosition_Internal(int entityID, glm::vec3* vector);
+			static bool HasComponent_Interal(uint64_t entityID,MonoObject* type);
+			static void SetPosition_Internal(uint64_t entityID, glm::vec3* vector);
+			static void GetPosition_Internal(uint64_t entityID, glm::vec3* vector);
 			static bool MouseKey_Internal(int button);
 			static void MousePosition_Internal(glm::vec2* pos);
 			static bool KeyPressed_Internal(uint16_t button);
